+++ pkg/nodeconfig/init.go
