+++ pkg/nodeconfig/network_test.go
