+++ test/e2e/metrics_test.go
@@ -21,7 +21,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 
-	"github.com/openshift/windows-machine-config-operator/pkg/controller/windowsmachine/metrics"
+	"github.com/openshift/windows-machine-config-operator/pkg/metrics"
 )
 
 const (
