+++ test/e2e/reconfigure_test.go
@@ -12,7 +12,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 
-	nc "github.com/openshift/windows-machine-config-operator/pkg/controller/windowsmachine/nodeconfig"
+	nc "github.com/openshift/windows-machine-config-operator/pkg/nodeconfig"
 )
 
 // reconfigurationTest tests that the correct behavior occurs when a previously configured Machine is configured
