+++ test/e2e/upgrade_test.go
@@ -16,7 +16,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/wait"
 
-	nc "github.com/openshift/windows-machine-config-operator/pkg/controller/windowsmachine/nodeconfig"
+	nc "github.com/openshift/windows-machine-config-operator/pkg/nodeconfig"
 )
 
 const (
